Switched to module augmentation of Message from Discord.js
