+++ src/message-features/BotSay.ts
@@ -1,9 +1,9 @@
 import { BaseHandler, MessageHandler } from "./BaseHandler";
 import { Commands } from "./Commands";
-import { MessageUtil } from "../MessageUtil";
+import { Message } from "discord.js";
 
 export class BotSay extends BaseHandler implements MessageHandler {
-  handleMessage(message: MessageUtil): void {
+  handleMessage(message: Message): void {
     // Responds with exactly what the previous user said
     if (message.commandIs(Commands.BotSay) && message.isAdmin()) {
       let messageText: string = message.toString(true);
