+++ src/message-features/LockChannel.ts
@@ -1,9 +1,9 @@
 import { BaseHandler, MessageHandler } from "./BaseHandler";
+import { Message } from "discord.js";
 import { Commands } from "./Commands";
-import { MessageUtil } from "../MessageUtil";
 
 export class LockChannel extends BaseHandler implements MessageHandler {
-  handleMessage(message: MessageUtil): void {
+  handleMessage(message: Message): void {
     if (message.commandIs(Commands.LockChannel)) {
       this.util.toggleChannelLock();
     }
