+++ src/message-features/PingPong.ts
@@ -1,5 +1,5 @@
+import { Message } from "discord.js";
 import { BaseHandler, MessageHandler } from "./BaseHandler";
-import { MessageUtil } from "../MessageUtil";
 
 enum Keywords {
   trigger = "ping",
@@ -7,7 +7,7 @@ enum Keywords {
 }
 
 export class PingPong extends BaseHandler implements MessageHandler {
-  handleMessage(message: MessageUtil): void {
+  handleMessage(message: Message): void {
     if (message.toString() === Keywords.trigger) {
       message.sendToChannel(Keywords.response);
       console.log(`Responding ${Keywords.response} because user said ${Keywords.trigger}`);
